moved all message modal components into a separate folder
